# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Renovo API
  version: "1.0.0"
  description: |
    Read and write subscriptions, categories, payment methods, tags and attachments.

    **Authentication.** Every endpoint except the two OpenAPI documents needs a
    token, issued under Settings → API tokens and sent as
    `Authorization: Bearer rnv_<public id>_<secret>`. Session cookies are not
    accepted: the API is deliberately unreachable with a browser's ambient
    credentials, which is what makes it safe for it to be exempt from CSRF
    tokens.

    **A token narrows, it never grants.** The request runs under the household
    role of the account that issued the token, so a Viewer's write token still
    cannot write. A read-only token is refused any unsafe method outright.

    **Money is always an integer in minor units** (`price_minor: 1999` with
    `currency: GBP` is £19.99), never a decimal. JSON numbers are binary floats
    in most clients and 19.99 is not representable in one.

    **There is no PATCH.** `PUT` replaces the whole resource. To change one
    field, `GET` the resource, edit the representation and `PUT` it back.

    **Not in version 1.** Shared-cost splits, scheduled price changes and the
    usage counter are editable in the web interface but have no endpoints here.
    Each is a sub-resource with rules of its own — a split has to total its
    shares, a scheduled price has an effective date that interacts with the
...
  license:
    name: MIT
servers:
  - url: /
...
security:
  - bearerToken: []
paths:
...
  /api/v1/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions.
      description: |
        Returns only what the token's role and the instance's isolation mode
        allow — the same rows the same user would see in the web interface.
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Search names and notes. }
        - { name: category, in: query, schema: { type: integer }, description: Category id. }
        - { name: owner, in: query, schema: { type: integer }, description: Owner user id. }
        - { name: currency, in: query, schema: { type: string }, description: ISO 4217 code. }
        - name: type
          in: query
          schema: { type: string, enum: [recurring, one_off, lifetime] }
        - name: inactive
          in: query
          schema: { type: string, enum: ["1"] }
          description: Include inactive subscriptions — paused and cancelled.
        - name: status
          in: query
          schema: { type: string, enum: [active, trial, paused, cancelled] }
          description: >
            Only subscriptions in this state (see `status` on the resource).
            Takes precedence over `inactive`.
        - name: sort
          in: query
          schema: { type: string, enum: [name, next_payment, price, created] }
          description: >
            Whatever is chosen, inactive subscriptions are ordered after active
            ones, and rows with no value in the sort column (a lifetime licence
            has no next payment date) after those that have one.
        - { name: dir, in: query, schema: { type: string, enum: [asc, desc] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: A page of subscriptions.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subscription"
                  meta:
                    $ref: "#/components/schemas/Page"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Subscriptions]
      summary: Create a subscription.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created.
          headers:
            Location:
              schema: { type: string }
              description: The new subscription's URL.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Subscription"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/subscriptions/{id}:
    parameters:
      - $ref: "#/components/parameters/SubscriptionId"
    get:
      tags: [Subscriptions]
      summary: Fetch one subscription.
      responses:
        "200":
          description: The subscription.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Subscription"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Subscriptions]
      summary: Replace a subscription.
      description: |
        A full replacement. Any field the body omits takes its documented
        default, so read the resource and send it back edited rather than
        sending only what changed.

        The logo is the one exception: it is a file with endpoints of its own,
        and a `PUT` leaves it as it is.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "200":
          description: The updated subscription.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Subscription"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"

    delete:
      tags: [Subscriptions]
      summary: Delete a subscription.
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: An API token issued under Settings → API tokens.
  parameters:
    SubscriptionId:
      name: id
      in: path
      required: true
      schema: { type: integer }
  responses:
...
    Unauthorized:
      description: No token was presented, or it is revoked, expired or unknown.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The token's role does not permit this.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        No such row — or one the caller may not see. The two are deliberately
        indistinguishable, so that probing for another household's ids tells the
        prober nothing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: The body was understood but is not valid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [status, title, message]
          properties:
            status: { type: integer }
            title: { type: string }
            message: { type: string }
            errors:
              type: object
              description: Field name to message. Present on a 422.
              additionalProperties: { type: string }
    Page:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
...
    Subscription:
      allOf:
        - $ref: "#/components/schemas/SubscriptionInput"
        - type: object
          required:
            - id
            - price_minor
            - currency
            - is_active
            - created_at
            - updated_at
          properties:
            id: { type: integer }
            anchor_day:
              type: integer
              nullable: true
              description: The day of the month billing returns to after a short month.
            category_name: { type: string, nullable: true }
            payment_method_name: { type: string, nullable: true }
            split_mode: { type: string }
            usage_count: { type: integer }
            usage_rating: { type: integer, nullable: true }
            logo_path:
              type: string
              nullable: true
              description: Read-only. Set it with the logo endpoints.
            website_url:
              type: string
              nullable: true
              description: |
                Where the subscription is managed. Also the domain its icon is
                fetched from when no logo has been uploaded — that fetch goes
                through the instance's guarded HTTP client and is cached per
                domain.
            monthly_minor:
              type: integer
              nullable: true
              description: Read-only. Equivalent monthly cost in minor units.
            yearly_minor: { type: integer, nullable: true }
            next_charge_date:
              type: string
              format: date
              nullable: true
              description: |
                Read-only. The trial's conversion date while a trial is running,
                and the next payment date otherwise.
            cancellation_deadline:
              type: string
              format: date
              nullable: true
              description: Read-only. The last day to cancel and avoid the next charge.
            days_until_next_payment: { type: integer, nullable: true }
            cancelled_at:
              type: string
              format: date
              nullable: true
              description: |
                Read-only. The day it was cancelled. Set and cleared by the
                cancel and uncancel endpoints only.
            status:
              type: string
              enum: [active, trial, paused, cancelled]
              description: |
                Read-only. Derived in this order: cancelled, then paused
                (inactive and not cancelled), then trial, then active.
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }